week6/requirements.txt
fastapi
uvicorn[standard]
jinja2
mysql-connector-python